initial commit
basically just trying to solve the first problem, and my approach. at this point we can take an input (float) balance, and calculate a few parameters. Then, we can iterate over each month and show the remaining balance. The calculation is not fully set up, so the output is incorrect (I still need to take into account all the various rates, etc), however the idea of taking the balance and iterating for each month and showing the output is there..